+++ R/spatialSPsurv.R
@@ -5,16 +5,21 @@
 #' @param immune split stage equation written in a formula of the form C ~ Z1 + Z2 + ... where C is a binary indicator of immunity.
 #' @param Y0 the elapsed time since inception until the beginning of time period (t-1).
 #' @param LY last observation year (coded as 1; 0 otherwise) due to censoring or failure.
-#' @param data data.frame.
 #' @param S spatial information (e.g. district ID) for each observation that matches the spatial matrix row/column information.
 #' @param A an a times a spatial weights matrix where a is the number of unique spatial units (S) load as a separate file.
+#' @param data data.frame.
 #' @param N number of MCMC iterations.
 #' @param burn burn-in to be discarded.
 #' @param thin thinning to prevent from autocorrelation.
 #' @param w size of the slice in the slice sampling for (betas, gammas, rho). Write it as a vector. E.g. c(1,1,1).
 #' @param m limit on steps in the slice sampling. A vector of values for beta, gamma, rho.
+#' @param ini.beta initial value for the parameter vector beta.  By default is 0.
+#' @param ini.gamma initial value for the parameter vector gamma. By default is 0.
+#' @param ini.W initial value for the parameter vector W. By default is 0.
+#' @param ini.V initial value for the parameter vector V.  By default is 0.
 #' @param form type of parametric model (Weibull, Exponential, or Log-Logistic).
-#' @param prop.var proposal variance for Metropolis-Hastings.
+#' @param prop.varV proposal for variance of V in Metropolis-Hastings.
+#' @param prop.varW proposal for variance of W in Metropolis-Hastings.
 #' @param id_WV vector of type character that modifies the colnames of W and V in the model’s result. By default is \code{colnames(A)}.
 #'
 #' @return spatialSPsurv returns an object of class \code{"spatialSPsurv"}.
@@ -33,6 +38,10 @@
 #' \item{Y0}{vector of `Y0'.}
 #' \item{C}{vector of `C'.}
 #' \item{S}{vector of `S'.}
+#' \item{ini.beta}{numeric initial values of beta.}
+#' \item{ini.gamma}{numeric initial values of gamma.}
+#' \item{ini.W}{numeric initial values of W.}
+#' \item{ini.V}{numeric initial values of V.}
 #' \item{form}{character, type of distribution.}
 #' \item{call}{description for the model to be estimated.}
 #'
@@ -48,21 +57,22 @@
 #'
 #' model <-
 #'     spatialSPsurv(
-#'         duration = duration ~ fhcompor1 + lgdpl + comprehensive + victory +
-#'                    instabl + intensityln + ethfrac + unpko,
-#'         immune   = cured ~ fhcompor1 + lgdpl + victory,
-#'         Y0       = 't.0',
-#'         LY       = 'lastyear',
-#'         S        = 'sp_id' ,
-#'         data     = walter[[1]],
-#'         N        = 100,
-#'         burn     = 10,
-#'         thin     = 10,
-#'         w        = c(1,1,1),
-#'         m        = 10,
-#'         form     = "Weibull",
-#'         prop.var = 1e-05,
-#'         A        = walter[[2]]
+#'         duration  = duration ~ fhcompor1 + lgdpl + comprehensive + victory +
+#'                     instabl + intensityln + ethfrac + unpko,
+#'         immune    = cured ~ fhcompor1 + lgdpl + victory,
+#'         Y0        = 't.0',
+#'         LY        = 'lastyear',
+#'         S         = 'sp_id' ,
+#'         data      = walter[[1]],
+#'         N         = 100,
+#'         burn      = 10,
+#'         thin      = 10,
+#'         w         = c(1,1,1),
+#'         m         = 10,
+#'         form      = "Weibull",
+#'         prop.varV = 1e-05,
+#'         prop.varW = 1e-05,
+#'         A         = walter[[2]]
 #'     )
 #'
 #' print(model)
@@ -87,8 +97,13 @@ spatialSPsurv <- function(duration,
                          thin,
                          w = c(1, 1, 1),
                          m = 10,
+                         ini.beta =  0,
+                         ini.gamma = 0,
+                         ini.W = 0,
+                         ini.V= 0,
                          form = c('Weibull', 'exponential', 'loglog'),
-                         prop.var,
+                         prop.varV,
+                         prop.varW,
                          id_WV = colnames(A))
 {
 
@@ -96,18 +111,25 @@ spatialSPsurv <- function(duration,
     model <- 'spatialSPsurv'
     r   <- formcall(duration = duration, immune = immune, data = data, Y0 = Y0,
                     LY = LY, S = S, N = N, burn = burn, thin = thin, w = w, m = m,
-                    form = dis, prop.var = prop.var, A = A, model = 'spatialSPsurv')
+                    ini.beta = ini.beta, ini.gamma = ini.gamma,
+                    ini.W = ini.W, ini.V = ini.V,
+                    form = dis, prop.varV = prop.varV, prop.varW = prop.varW,
+                    A = A, model = 'spatialSPsurv')
 
     if(form == 'loglog') {
         results <- mcmcSpatialLog(Y = r$Y, Y0 = r$Y0, C = r$C, LY = r$LY, X = r$X, Z = r$Z,
                                  S = r$S, N = r$N, burn = r$burn, thin = r$thin, w = r$w,
-                                 m = r$m, form = r$form, prop.var = r$prop.var, A = r$A,
-                                 id_WV = id_WV)
+                                 m = r$m, ini.beta = r$ini.beta, ini.gamma = r$ini.gamma,
+                                 ini.W = r$ini.W, ini.V = r$ini.V,
+                                 form = r$form, prop.varV = r$prop.varV, prop.varW = r$prop.varW,
+                                 A = r$A, id_WV = id_WV)
     } else {
         results <- mcmcspatialSP(Y = r$Y, Y0 = r$Y0, C = r$C, LY = r$LY, X = r$X, Z = r$Z,
                              S = r$S, N = r$N, burn = r$burn, thin = r$thin, w = r$w,
-                             m = r$m, form = r$form, prop.var = r$prop.var, A = r$A,
-                             id_WV = id_WV)
+                             m = r$m, ini.beta = r$ini.beta, ini.gamma = r$ini.gamma,
+                             ini.W = r$ini.W, ini.V = r$ini.V,
+                             form = r$form, prop.varV = r$prop.varV, prop.varW = r$prop.varW,
+                             A = r$A, id_WV = id_WV)
     }
     results$call   <- match.call()
     class(results) <- model
@@ -155,7 +177,7 @@ print.spatialSPsurv <- function(x, ...){
     cat('Duration equation: \n')
     print(summary(x, parameter = 'betas')$statistics)
     cat('\n')
-    cat('Inmune equation: \n')
+    cat('Immune equation: \n')
     print(summary(x, parameter = 'gammas')$statistics)
     cat('\n')