+++ NEWS.md
@@ -1,4 +1,11 @@
 
+
+# BayesSPsurv 0.1.3
+
+* parameter ini.gamma, ini.gamma, ini.W and ini.V   are added to `spatialSPsurv()`, `exchangeSPsurv()` and `pooledSPsurv()`
+* parameter prop.varV and prop.varW are added to `spatialSPsurv()` and `exchangeSPsurv()`
+* progress bar in `spatialSPsurv()`, `exchangeSPsurv()` and `pooledSPsurv()`
+
 # BayesSPsurv 0.1.2
 
 * threshold argument added to `plot_Moran.I()` and `plot_JoinCount()`
@@ -8,7 +15,4 @@
 # BayesSPsurv 0.1.1
 
 * add `plot_Moran.I()` and `plot_JointCount()`
-
-
-